Rename gdb interface and driver submodules
